Started moving things to modules
Moved functionality to modules where it made sense. The whole thing is
not hookep up yet, so it won't run (but will compile). Also moved input
validation into the library and added a Config struct instead of
requiring a ton of parameters.
